Separated QR and SVD and reworked QR routines to avoid code repetition in python
